拆分配置变更模块
